refactor(client-worker): add favorite workers and travel time config screens
This commit introduces two new screens: the Favorite Workers Screen
and the Travel Time Configuration Screen. The Favorite Workers Screen
allows users to view and manage their favorite workers, including
search functionality and the ability to remove favorites. The Travel
Time Configuration Screen enables users to set and save travel time
for appointments, with validation and feedback through snackbars.

These additions enhance the user experience by providing more
functionality related to worker management and scheduling.
