+++ composeApp/src/commonMain/kotlin/com/example/seviya/app/navigation/AppNavGraph.kt
@@ -25,19 +25,19 @@ import com.example.seviya.feature.client.ClientLocationCatalogScreen
 import com.example.seviya.feature.client.ClientMapScreen
 import com.example.seviya.feature.client.ClientPaymentUploadScreen
 import com.example.seviya.feature.client.ClientToWorkerReviewScreen
-import com.example.seviya.UI.DailyAgendaScreen
-import com.example.seviya.UI.FavoriteWorkersRoute
-import com.example.seviya.UI.MonthlyCalendarScreen
-import com.example.seviya.UI.ProfessionalProfileRoute
-import com.example.seviya.UI.RequestAppointmentRoute
+import com.example.seviya.feature.shared.DailyAgendaScreen
+import com.example.seviya.feature.worker.FavoriteWorkersRoute
+import com.example.seviya.feature.shared.MonthlyCalendarScreen
+import com.example.seviya.feature.shared.ProfessionalProfileRoute
+import com.example.seviya.feature.shared.RequestAppointmentRoute
 import com.example.seviya.feature.landing.RoleAdmissionCatalogScreen
 import com.example.seviya.feature.landing.RoleCatalogScreen
-import com.example.seviya.UI.TravelTimeConfigRoute
-import com.example.seviya.UI.WeeklyAgendaScreen
+import com.example.seviya.feature.worker.TravelTimeConfigRoute
+import com.example.seviya.feature.shared.WeeklyAgendaScreen
 import com.example.seviya.feature.worker.WorkerAppointmentDetailScreen
 import com.example.seviya.feature.worker.WorkerCategoriesRoute
 import com.example.seviya.feature.worker.WorkerDailyAppointmentsScreen
-import com.example.seviya.UI.WorkerDashboardRoute
+import com.example.seviya.feature.worker.WorkerDashboardRoute
 import com.example.seviya.feature.worker.WorkerPaymentDetailScreen
 import com.example.seviya.feature.worker.WorkerRequestDetailScreen
 import com.example.seviya.feature.worker.WorkerStartAppointmentOtpScreen
