+++ composeApp/src/commonMain/kotlin/com/example/seviya/feature/worker/TravelTimeConfigScreen.kt
@@ -1,4 +1,4 @@
-package com.example.seviya.UI
+package com.example.seviya.feature.worker
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.core.FastOutSlowInEasing
